ESP32-HAL-UART: Fixed hang (and WDT reboot) in UART int handler
`HardwareSerial::begin()` hangs in situation when the intensive RX stream exposed to UART port BEFORE `Serial.begin()`.
In this case Interrupt Handler execution takes a long time and WDT reboots the ESP32 with message: "Core  0 panic'ed (Interrupt wdt timeout on CPU0)"

This problem sometimes is not appear at all. But in most cases it appears during the ESP boot in the `setup()`.